+++ src/otx/algorithms/classification/adapters/mmcls/nncf/builder.py
@@ -123,6 +123,8 @@ def build_nncf_classifier(  # pylint: disable=too-many-locals,too-many-statement
         )
         state_dict = None
 
+    if config.data.test.pipeline[0]["type"] == "LoadImageFromOTXDataset":
+        config.data.test.pipeline = config.data.test.pipeline[1:]
     test_pipeline = Compose(config.data.test.pipeline)
     get_fake_input_fn = partial(
         get_fake_input,