+++ src/otx/algorithms/segmentation/utils/__init__.py
@@ -5,5 +5,6 @@
 
 
 from .metadata import get_seg_model_api_configuration
+from .processing import get_activation_map
 
-__all__ = ["get_seg_model_api_configuration"]
+__all__ = ["get_activation_map", "get_seg_model_api_configuration"]