+++ tests/unit/algorithms/detection/adapters/mmdet/test_configurer.py
@@ -3,6 +3,7 @@ import os
 
 import pytest
 import tempfile
+from mmcv.runner import CheckpointLoader
 from mmcv.utils import ConfigDict
 
 from otx.api.entities.model_template import TaskType
@@ -13,6 +14,7 @@ from otx.algorithms.detection.adapters.mmdet.configurer import (
     IncrDetectionConfigurer,
     SemiSLDetectionConfigurer,
 )
+from otx.algorithms.common.configs.configuration_enums import InputSizePreset
 from tests.test_suite.e2e_test_system import e2e_pytest_unit
 from tests.unit.algorithms.detection.test_helpers import (
     DEFAULT_DET_TEMPLATE_DIR,
@@ -23,11 +25,21 @@ from tests.unit.algorithms.detection.test_helpers import (
 class TestDetectionConfigurer:
     @pytest.fixture(autouse=True)
     def setup(self) -> None:
-        self.configurer = DetectionConfigurer()
+        self.configurer = DetectionConfigurer("detection", True)
         self.model_cfg = MPAConfig.fromfile(os.path.join(DEFAULT_DET_TEMPLATE_DIR, "model.py"))
-        self.data_cfg = MPAConfig.fromfile(os.path.join(DEFAULT_DET_TEMPLATE_DIR, "data_pipeline.py"))
-        self.model_cfg.merge_from_dict(self.data_cfg)
+        data_pipeline_cfg = MPAConfig.fromfile(os.path.join(DEFAULT_DET_TEMPLATE_DIR, "data_pipeline.py"))
+        self.model_cfg.merge_from_dict(data_pipeline_cfg)
+
         self.det_dataset, self.det_labels = generate_det_dataset(TaskType.DETECTION, 100)
+        self.data_cfg = ConfigDict(
+            {
+                "data": {
+                    "train": {"otx_dataset": self.det_dataset, "labels": self.det_labels},
+                    "val": {"otx_dataset": self.det_dataset, "labels": self.det_labels},
+                    "test": {"otx_dataset": self.det_dataset, "labels": self.det_labels},
+                }
+            }
+        )
 
     @e2e_pytest_unit
     def test_configure(self, mocker):
@@ -39,33 +51,30 @@ class TestDetectionConfigurer:
         mock_cfg_task = mocker.patch.object(DetectionConfigurer, "configure_task")
         mock_cfg_hook = mocker.patch.object(DetectionConfigurer, "configure_hook")
         mock_cfg_gpu = mocker.patch.object(DetectionConfigurer, "configure_samples_per_gpu")
-        mock_cfg_fp16_optimizer = mocker.patch.object(DetectionConfigurer, "configure_fp16_optimizer")
+        mock_cfg_fp16 = mocker.patch.object(DetectionConfigurer, "configure_fp16")
         mock_cfg_compat_cfg = mocker.patch.object(DetectionConfigurer, "configure_compat_cfg")
+        mock_cfg_input_size = mocker.patch.object(DetectionConfigurer, "configure_input_size")
 
         model_cfg = copy.deepcopy(self.model_cfg)
         data_cfg = copy.deepcopy(self.data_cfg)
-        returned_value = self.configurer.configure(model_cfg, self.det_dataset, "", data_cfg, True)
+        returned_value = self.configurer.configure(model_cfg, self.det_dataset, "", data_cfg)
         mock_cfg_base.assert_called_once_with(model_cfg, data_cfg, None, None)
-        mock_cfg_device.assert_called_once_with(model_cfg, True)
+        mock_cfg_device.assert_called_once_with(model_cfg)
         mock_cfg_model.assert_called_once_with(model_cfg, None)
         mock_cfg_ckpt.assert_called_once_with(model_cfg, "")
-        mock_cfg_regularization.assert_called_once_with(model_cfg, True)
-        mock_cfg_task.assert_called_once_with(model_cfg, self.det_dataset, True)
+        mock_cfg_regularization.assert_called_once_with(model_cfg)
+        mock_cfg_task.assert_called_once_with(model_cfg, self.det_dataset)
         mock_cfg_hook.assert_called_once_with(model_cfg)
         mock_cfg_gpu.assert_called_once_with(model_cfg, "train")
-        mock_cfg_fp16_optimizer.assert_called_once_with(model_cfg)
+        mock_cfg_fp16.assert_called_once_with(model_cfg)
         mock_cfg_compat_cfg.assert_called_once_with(model_cfg)
+        mock_cfg_input_size.assert_called_once_with(model_cfg, InputSizePreset.DEFAULT, "")
         assert returned_value == model_cfg
 
     @e2e_pytest_unit
     def test_configure_base(self, mocker):
-        mocker.patch(
-            "otx.algorithms.detection.adapters.mmdet.configurer.align_data_config_with_recipe",
-            return_value=True,
-        )
-
         model_cfg = copy.deepcopy(self.model_cfg)
-        data_cfg = copy.deepcopy(self.data_cfg._cfg_dict)
+        data_cfg = copy.deepcopy(self.data_cfg)
         self.configurer.configure_base(model_cfg, data_cfg, [], [])
 
     @e2e_pytest_unit
@@ -74,12 +83,12 @@ class TestDetectionConfigurer:
             "torch.distributed.is_initialized",
             return_value=True,
         )
+        mocker.patch("torch.distributed.get_world_size", return_value=2)
         world_size = 2
-        mocker.patch.object(configurer, "dist").get_world_size.return_value = world_size
         mocker.patch("os.environ", return_value={"LOCAL_RANK": 2})
         config = copy.deepcopy(self.model_cfg)
         origin_lr = config.optimizer.lr
-        self.configurer.configure_device(config, True)
+        self.configurer.configure_device(config)
         assert config.distributed is True
         assert config.optimizer.lr == pytest.approx(origin_lr * world_size)
 
@@ -92,7 +101,7 @@ class TestDetectionConfigurer:
             return_value=False,
         )
         config = copy.deepcopy(self.model_cfg)
-        self.configurer.configure_device(config, True)
+        self.configurer.configure_device(config)
         assert config.distributed is False
         assert config.device == "cpu"
 
@@ -105,7 +114,7 @@ class TestDetectionConfigurer:
             return_value=True,
         )
         config = copy.deepcopy(self.model_cfg)
-        self.configurer.configure_device(config, True)
+        self.configurer.configure_device(config)
         assert config.distributed is False
         assert config.device == "cuda"
 
@@ -120,8 +129,9 @@ class TestDetectionConfigurer:
         model_cfg = copy.deepcopy(self.model_cfg)
         model_cfg.resume = True
 
-        mocker.patch(
-            "otx.algorithms.detection.adapters.mmdet.configurer.CheckpointLoader.load_checkpoint",
+        mocker.patch.object(
+            CheckpointLoader,
+            "load_checkpoint",
             return_value={"model": None},
         )
         with tempfile.TemporaryDirectory() as tempdir:
@@ -165,7 +175,7 @@ class TestDetectionConfigurer:
                 ],
             ),
         )
-        self.configurer.configure_data(self.model_cfg, True, data_cfg)
+        self.configurer.configure_data(self.model_cfg, data_cfg)
         assert self.model_cfg.data
         assert self.model_cfg.data.train
         assert self.model_cfg.data.val
@@ -174,23 +184,24 @@ class TestDetectionConfigurer:
     def test_configure_task(self, mocker):
         ssd_dir = os.path.join("src/otx/algorithms/detection/configs/detection", "mobilenetv2_ssd")
         ssd_cfg = MPAConfig.fromfile(os.path.join(ssd_dir, "model.py"))
+        data_pipeline_cfg = MPAConfig.fromfile(os.path.join(ssd_dir, "data_pipeline.py"))
         ssd_cfg.task_adapt = {"type": "mpa", "op": "REPLACE", "use_mpa_anchor": True}
         model_cfg = copy.deepcopy(ssd_cfg)
-        model_cfg.merge_from_dict(self.data_cfg)
-        self.configurer.configure_task(model_cfg, self.det_dataset, True)
+        model_cfg.merge_from_dict(data_pipeline_cfg)
+        self.configurer.configure_task(model_cfg, self.det_dataset)
         assert model_cfg.model.bbox_head.anchor_generator != ssd_cfg.model.bbox_head.anchor_generator
 
         model_cfg = copy.deepcopy(self.model_cfg)
         model_cfg.task_adapt = {"type": "mpa", "op": "REPLACE", "use_mpa_anchor": True}
         model_cfg.model.bbox_head.type = "ATSSHead"
-        self.configurer.configure_task(model_cfg, self.det_dataset, True)
+        self.configurer.configure_task(model_cfg, self.det_dataset)
 
         model_cfg.model.bbox_head.type = "VFNetHead"
-        self.configurer.configure_task(model_cfg, self.det_dataset, True)
+        self.configurer.configure_task(model_cfg, self.det_dataset)
 
         model_cfg.model.bbox_head.type = "YOLOXHead"
         model_cfg.data.train.type = "MultiImageMixDataset"
-        self.configurer.configure_task(model_cfg, self.det_dataset, True)
+        self.configurer.configure_task(model_cfg, self.det_dataset)
 
         def mock_configure_classes(*args, **kwargs):
             return True
@@ -199,7 +210,7 @@ class TestDetectionConfigurer:
         self.configurer.model_classes = []
         self.configurer.data_classes = ["red", "green"]
         self.configurer.configure_classes = mock_configure_classes
-        self.configurer.configure_task(model_cfg, self.det_dataset, True)
+        self.configurer.configure_task(model_cfg, self.det_dataset)
 
     @e2e_pytest_unit
     def test_configure_hook(self):
@@ -216,20 +227,20 @@ class TestDetectionConfigurer:
         assert model_cfg.data.train_dataloader == {"samples_per_gpu": 1, "drop_last": True}
 
     @e2e_pytest_unit
-    def test_configure_fp16_optimizer(self):
+    def test_configure_fp16(self):
         model_cfg = copy.deepcopy(self.model_cfg)
         model_cfg.fp16 = {}
-        self.configurer.configure_fp16_optimizer(model_cfg)
+        self.configurer.configure_fp16(model_cfg)
         assert model_cfg.optimizer_config.type == "Fp16OptimizerHook"
 
         model_cfg.fp16 = {}
         model_cfg.optimizer_config.type = "SAMOptimizerHook"
-        self.configurer.configure_fp16_optimizer(model_cfg)
+        self.configurer.configure_fp16(model_cfg)
         assert model_cfg.optimizer_config.type == "Fp16SAMOptimizerHook"
 
         model_cfg.fp16 = {}
         model_cfg.optimizer_config.type = "DummyOptimizerHook"
-        self.configurer.configure_fp16_optimizer(model_cfg)
+        self.configurer.configure_fp16(model_cfg)
         assert model_cfg.optimizer_config.type == "DummyOptimizerHook"
 
     @e2e_pytest_unit
@@ -244,7 +255,7 @@ class TestDetectionConfigurer:
     def test_configure_regularization(self):
         configure_cfg = copy.deepcopy(self.model_cfg)
         configure_cfg.model.l2sp_weight = 1.0
-        self.configurer.configure_regularization(configure_cfg, True)
+        self.configurer.configure_regularization(configure_cfg)
         assert "l2sp_ckpt" in configure_cfg.model
         assert configure_cfg.optimizer.weight_decay == 0.0
 
@@ -254,11 +265,65 @@ class TestDetectionConfigurer:
         config.data.train.dataset = ConfigDict({"dataset": [1, 2, 3]})
         assert [1, 2, 3] == self.configurer.get_data_cfg(config, "train")
 
+    @e2e_pytest_unit
+    @pytest.mark.parametrize("input_size", [None, (256, 256)])
+    def test_configure_input_size_not_yolox(self, mocker, input_size):
+        # prepare
+        mock_cfg = mocker.MagicMock()
+        mocker.patch.object(configurer, "get_configured_input_size", return_value=input_size)
+        mock_input_manager = mocker.MagicMock()
+        mock_input_manager_cls = mocker.patch.object(configurer, "InputSizeManager")
+        mock_input_manager_cls.return_value = mock_input_manager
+
+        # excute
+        self.configurer.configure_input_size(mock_cfg, InputSizePreset.DEFAULT, self.data_cfg)
+
+        # check
+        if input_size is not None:
+            mock_input_manager_cls.assert_called_once_with(mock_cfg.data, None)
+            mock_input_manager.set_input_size.assert_called_once_with(input_size)
+        else:
+            mock_input_manager_cls.assert_not_called()
+
+    @e2e_pytest_unit
+    @pytest.mark.parametrize("input_size", [(256, 256), (300, 300)])
+    @pytest.mark.parametrize("is_yolox_tiny", [True, False])
+    def test_configure_input_size_yolox(self, mocker, input_size, is_yolox_tiny):
+        # prepare
+        mock_cfg = mocker.MagicMock()
+        mock_cfg.model.type = "CustomYOLOX"
+        if is_yolox_tiny:
+            mock_cfg.model.backbone.widen_factor = 0.375
+            base_input_size = {
+                "train": 640,
+                "val": 416,
+                "test": 416,
+            }
+        else:
+            base_input_size = None
+        mocker.patch.object(configurer, "get_configured_input_size", return_value=input_size)
+        mock_input_manager = mocker.MagicMock()
+        mock_input_manager_cls = mocker.patch.object(configurer, "InputSizeManager")
+        mock_input_manager_cls.return_value = mock_input_manager
+
+        # If model is one of yolox variants and input size isn't multiple of 32, error should be raised.
+        if input_size[0] % 32 != 0:
+            with pytest.raises(ValueError):
+                self.configurer.configure_input_size(mock_cfg, InputSizePreset.DEFAULT, self.data_cfg)
+            return
+
+        # excute
+        self.configurer.configure_input_size(mock_cfg, InputSizePreset.DEFAULT, self.data_cfg)
+
+        # check
+        mock_input_manager_cls.assert_called_once_with(mock_cfg.data, base_input_size)
+        mock_input_manager.set_input_size.assert_called_once_with(input_size)
+
 
 class TestIncrDetectionConfigurer:
     @pytest.fixture(autouse=True)
     def setup(self) -> None:
-        self.configurer = IncrDetectionConfigurer()
+        self.configurer = IncrDetectionConfigurer("detection", True)
         self.model_cfg = MPAConfig.fromfile(os.path.join(DEFAULT_DET_TEMPLATE_DIR, "model.py"))
         self.data_cfg = MPAConfig.fromfile(os.path.join(DEFAULT_DET_TEMPLATE_DIR, "data_pipeline.py"))
         self.det_dataset, self.det_labels = generate_det_dataset(TaskType.DETECTION, 100)
@@ -267,7 +332,7 @@ class TestIncrDetectionConfigurer:
         mocker.patch.object(DetectionConfigurer, "configure_task")
         self.model_cfg.task_adapt = {}
         self.configurer.task_adapt_type = "mpa"
-        self.configurer.configure_task(self.model_cfg, self.det_dataset, True)
+        self.configurer.configure_task(self.model_cfg, self.det_dataset)
         assert self.model_cfg.custom_hooks[1].type == "TaskAdaptHook"
         assert self.model_cfg.custom_hooks[1].sampler_flag is False
 
@@ -275,7 +340,7 @@ class TestIncrDetectionConfigurer:
 class TestSemiSLDetectionConfigurer:
     @pytest.fixture(autouse=True)
     def setup(self) -> None:
-        self.configurer = SemiSLDetectionConfigurer()
+        self.configurer = SemiSLDetectionConfigurer("detection", True)
         self.model_cfg = MPAConfig.fromfile(os.path.join(DEFAULT_DET_TEMPLATE_DIR, "model.py"))
         self.data_cfg = MPAConfig.fromfile(os.path.join(DEFAULT_DET_TEMPLATE_DIR, "data_pipeline.py"))
         self.model_cfg.merge_from_dict(self.data_cfg)
@@ -288,11 +353,11 @@ class TestSemiSLDetectionConfigurer:
         self.model_cfg.data.unlabeled = ConfigDict({"type": "OTXDataset", "otx_dataset": range(10)})
         self.model_cfg.model_task = "detection"
         self.model_cfg.distributed = False
-        self.configurer.configure_data(self.model_cfg, True, self.data_cfg)
+        self.configurer.configure_data(self.model_cfg, self.data_cfg)
 
     def test_configure_task(self):
         self.model_cfg.task_adapt = {"type": "mpa", "op": "REPLACE", "use_mpa_anchor": True}
-        self.configurer.configure_task(self.model_cfg, self.det_dataset, True)
+        self.configurer.configure_task(self.model_cfg, self.det_dataset)
 
         self.model_cfg.task_adapt = {"type": "not_mpa", "op": "REPLACE", "use_mpa_anchor": True}
-        self.configurer.configure_task(self.model_cfg, self.det_dataset, True)
+        self.configurer.configure_task(self.model_cfg, self.det_dataset)